Move entity extenions to domain project / remove IProductAttributeParser
