+++ src/Business/Grand.Business.Catalog/Services/Products/StockQuantityService.cs
@@ -1,4 +1,5 @@
-﻿using Grand.Business.Core.Interfaces.Catalog.Products;
+﻿using Grand.Business.Core.Extensions;
+using Grand.Business.Core.Interfaces.Catalog.Products;
 using Grand.Business.Core.Interfaces.Common.Localization;
 using Grand.Domain.Catalog;
 using Grand.Domain.Common;
@@ -8,14 +9,10 @@ namespace Grand.Business.Catalog.Services.Products
     public class StockQuantityService : IStockQuantityService
     {
         private readonly ITranslationService _translationService;
-        private readonly IProductAttributeParser _productAttributeParser;
-
         public StockQuantityService(
-            ITranslationService translationService,
-            IProductAttributeParser productAttributeParser)
+            ITranslationService translationService)
         {
             _translationService = translationService;
-            _productAttributeParser = productAttributeParser;
         }
 
         public virtual int GetTotalStockQuantity(Product product, bool useReservedQuantity = true,
@@ -154,7 +151,7 @@ namespace Grand.Business.Catalog.Services.Products
                         if (!product.StockAvailability)
                             return stockMessage;
 
-                        var combination = _productAttributeParser.FindProductAttributeCombination(product, attributes);
+                        var combination = product.FindProductAttributeCombination(attributes);
                         if (combination != null)
                         {
                             //combination exists
