--- src/Business/Grand.Business.Core/Extensions/BlogExtensions.cs
@@ -1,23 +1,0 @@
-using Grand.Domain.Blogs;
-
-namespace Grand.Business.Core.Extensions
-{
-    /// <summary>
-    /// Extensions
-    /// </summary>
-    public static class BlogExtensions
-    {
-        /// <summary>
-        /// Returns all posts published between the two dates.
-        /// </summary>
-        /// <param name="source">Source</param>
-        /// <param name="dateFrom">Date from</param>
-        /// <param name="dateTo">Date to</param>
-        /// <returns>Filtered posts</returns>
-        public static IList<BlogPost> GetPostsByDate(this IList<BlogPost> source,
-            DateTime dateFrom, DateTime dateTo)
-        {
-            return source.Where(p => dateFrom.Date <= (p.StartDateUtc ?? p.CreatedOnUtc) && (p.StartDateUtc ?? p.CreatedOnUtc).Date <= dateTo).ToList();
-        }
-    }
-}
