+++ src/Core/Grand.Domain/Common/AddressExtensions.cs
@@ -1,6 +1,4 @@
-using Grand.Domain.Common;
-
-namespace Grand.Business.Core.Extensions
+namespace Grand.Domain.Common
 {
     public static class AddressExtensions
     {
